moved create plant modal into its own component and updated create plant submit functionality
